+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## Fixes
 - Fixed a problem with Oracle Listener ports iteration.
+- Fixed a problem with the validation of the module args where depending on the ansible version an error could happen.
 
 # 1.0.1